cognisafe-deploy/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
sentence-transformers
spacy
scikit-learn
librosa
soundfile
numpy
scipy
torch
transformers